blurring and edge detection
use libraries to read and write files in the TIFF image format.  One
option in TIFF files is that they can be compressed if there are many
adjacent pixels of the same color; the compressed form is called a
run-length encoding; convert an image into a run-length encoding and
back.